requirements.txt
python-telegram-bot==13.15
mercadopago==2.1.0
